Move setup configuration to step itself
